Move BoundPatterns into their own class so that they can be shared.
